Add memory view mappings
Allows to pass memory views directly to JS (typed arrays on JS side will point into same address in the heap, just with correct offset/length, so conversion is very cheap).

When passing Box<[_]> or Vec<_> and not &'static [_], JS will have to know about this and take care of freeing it on its own.